Fixed signal parsing, improved debug
* rpdthread.h, rpdthread.cpp(Line 172 of old file): Added SHARED_MEM_SIZE
constant.
* rpdthread.cpp(71-152): Reworked the signal parsing: before this
commit, the signal went straight through the appropriate part of the
switch, wich had to redirect it if a second instruction was found,
however often instructions were missed because not redirected. In this
commit I introduced a QStringList wich collects all the strings of the
signal and then a for loop cycles on it until all the instructions have
been consumed. This way no instruction can be lost.
* rpdthread.cpp(85-94): Updated the filters for a valid clocks data path
(security + existance + not being a folder).
* rpdthread.cpp(119-130): Now SET_VALUE wants a separator between each
of its 3 strings (radeon-profile needs its last commit to work well with
it).
* rpdthread.cpp(132-141): Now TIMER_ON wants a separator between its
strings (radeon-profile needs its last commit to work well with it).
* rpdthread.cpp(155): Now readData tries to attach to shared memory if
it is not attached
* rpdthread.cpp(158-177): Before writing, the memory pointer is checked
to prevent segfaults, also QByteArray now manages most of the data
movements through QByteArray::constData and QFile::readAll.
* rpdthread.cpp(180-190): Updated the filters for a valid path to be set
with the value (security + existance + not being a folder).