Make UNIX socket group writeable
In Linux, UNIX sockets honor the permissions of the directory they are
in. Now that the UNIX socket was moved to its own subdirectory in /run,
which is not accessible by other, the socket is no longer world
writeable.

However, this is not enough to make the socket writeable by the group
that owns the subdirectory. Make the socket group writeable. This way,
it will be writeable by the group owning the directory the socket is in.

Closes marazmista/radeon-profile-daemon#18

Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>